refactor with mixin
