+++ src/views/StandaloneDirectorsFiling.vue
@@ -323,7 +323,7 @@ import { Certify, StaffPayment, SummaryDirectors, SummaryCertify, SummaryStaffPa
 import { ConfirmDialog, PaymentErrorDialog, ResumeErrorDialog, SaveErrorDialog, BcolErrorDialog } from '@/components/dialogs'
 
 // Mixins
-import { CommonMixin, FilingMixin, ResourceLookupMixin } from '@/mixins'
+import { CommonMixin, FilingMixin, ResourceLookupMixin, BcolMixin } from '@/mixins'
 
 // Enums and Constants
 import { EntityTypes, FilingCodes, FilingStatus, FilingTypes } from '@/enums'
@@ -348,7 +348,7 @@ export default {
     BcolErrorDialog
   },
 
-  mixins: [CommonMixin, FilingMixin, ResourceLookupMixin],
+  mixins: [CommonMixin, FilingMixin, ResourceLookupMixin, BcolMixin],
 
   data () {
     return {
@@ -639,14 +639,14 @@ export default {
           this.haveChanges = false
         }).catch(async error => {
           if (error && error.response && error.response.status === PAYMENT_REQUIRED) {
-            if (error.response.data && error.response.data.errors) {
-              const msgCode = error.response.data.errors.find(x => x.payment_error_type.startsWith('BCOL'))
-              if (msgCode) {
-                const errObj = await this.getErrorObj(msgCode.payment_error_type)
+            const errCode = this.getErrorCode(error)
+            if (errCode) {
+              const errObj = await this.getErrorObj(this.payApiUrl, errCode.payment_error_type)
+              if (errObj) {
                 this.bcolErrMsg = errObj.detail
                 this.bcolTitle = errObj.title
+                this.bcolErrorDialog = true
               }
-              this.bcolErrorDialog = true
             } else {
               this.paymentErrorDialog = true
             }
@@ -678,14 +678,14 @@ export default {
           this.haveChanges = false
         }).catch(async error => {
           if (error && error.response && error.response.status === PAYMENT_REQUIRED) {
-            if (error.response.data && error.response.data.errors) {
-              const msgCode = error.response.data.errors.find(x => x.payment_error_type.startsWith('BCOL'))
-              if (msgCode) {
-                const errObj = await this.getErrorObj(msgCode.payment_error_type)
+            const errCode = this.getErrorCode(error)
+            if (errCode) {
+              const errObj = await this.getErrorObj(this.payApiUrl, errCode.payment_error_type)
+              if (errObj) {
                 this.bcolErrMsg = errObj.detail
                 this.bcolTitle = errObj.title
+                this.bcolErrorDialog = true
               }
-              this.bcolErrorDialog = true
             } else {
               this.paymentErrorDialog = true
             }
@@ -703,13 +703,6 @@ export default {
         })
         return filing
       }
-    },
-
-    async getErrorObj (errCode) {
-      const fetchUrl = this.payApiUrl + 'codes/errors/BCOL_ERROR'
-      const errObj = await axios.get(fetchUrl)
-      console.log(errObj)
-      return errObj.data
     },
 
     navigateToDashboard (ignoreChanges: boolean = false) {
@@ -794,6 +787,7 @@ export default {
 
     resetErrors () {
       this.saveErrorDialog = false
+      this.bcolErrorDialog = false
       this.saveErrors = []
       this.saveWarnings = []
     },
