+++ src/views/StandaloneOfficeAddressFiling.vue
@@ -197,7 +197,7 @@ import { PAYMENT_REQUIRED, BAD_REQUEST } from 'http-status-codes'
 import { DASHBOARD } from '@/constants'
 
 // Mixins
-import { CommonMixin, FilingMixin, ResourceLookupMixin } from '@/mixins'
+import { CommonMixin, FilingMixin, ResourceLookupMixin, BcolMixin } from '@/mixins'
 
 // Enums
 import { EntityTypes, FilingCodes, FilingStatus, FilingTypes } from '@/enums'
@@ -216,7 +216,7 @@ export default {
     SaveErrorDialog,
     BcolErrorDialog
   },
-  mixins: [CommonMixin, FilingMixin, ResourceLookupMixin],
+  mixins: [CommonMixin, FilingMixin, ResourceLookupMixin, BcolMixin],
 
   data () {
     return {
@@ -609,14 +609,14 @@ export default {
           this.haveChanges = false
         }).catch(async error => {
           if (error && error.response && error.response.status === PAYMENT_REQUIRED) {
-            if (error.response.data && error.response.data.errors) {
-              const msgCode = error.response.data.errors.find(x => x.payment_error_type.startsWith('BCOL'))
-              if (msgCode) {
-                const errObj = await this.getErrorObj(msgCode.payment_error_type)
+            const errCode = this.getErrorCode(error)
+            if (errCode) {
+              const errObj = await this.getErrorObj(this.payApiUrl, errCode.payment_error_type)
+              if (errObj) {
                 this.bcolErrMsg = errObj.detail
                 this.bcolTitle = errObj.title
+                this.bcolErrorDialog = true
               }
-              this.bcolErrorDialog = true
             } else {
               this.paymentErrorDialog = true
             }
@@ -648,14 +648,14 @@ export default {
           this.haveChanges = false
         }).catch(async error => {
           if (error && error.response && error.response.status === PAYMENT_REQUIRED) {
-            if (error.response.data && error.response.data.errors) {
-              const msgCode = error.response.data.errors.find(x => x.payment_error_type.startsWith('BCOL'))
-              if (msgCode) {
-                const errObj = await this.getErrorObj(msgCode.payment_error_type)
+            const errCode = this.getErrorCode(error)
+            if (errCode) {
+              const errObj = await this.getErrorObj(this.payApiUrl, errCode.payment_error_type)
+              if (errObj) {
                 this.bcolErrMsg = errObj.detail
                 this.bcolTitle = errObj.title
+                this.bcolErrorDialog = true
               }
-              this.bcolErrorDialog = true
             } else {
               this.paymentErrorDialog = true
             }
@@ -675,13 +675,6 @@ export default {
       }
     },
 
-    async getErrorObj (errCode) {
-      const fetchUrl = this.payApiUrl + 'codes/errors/BCOL_ERROR'
-      const errObj = await axios.get(fetchUrl)
-      console.log(errObj)
-      return errObj.data
-    },
-
     navigateToDashboard (ignoreChanges: boolean = false) {
       if (ignoreChanges) this.haveChanges = false
       this.$router.push({ name: DASHBOARD })
@@ -689,6 +682,7 @@ export default {
 
     resetErrors () {
       this.saveErrorDialog = false
+      this.bcolErrorDialog = false
       this.saveErrors = []
       this.saveWarnings = []
     },
